requirements.txt
Pillow
requests
huggingface_hub
transformers
sentencepiece
protobuf
aiohttp
safetensors
gradio_client
boto3
groq
openai